Ensure a remap.config parameter exists
Assigning servers to Delivery Services cannot succeed unless some
Parameter - not necessarily assigned to any Cache Group or the Profile
of that or any other Delivery Service or server used or not used by that
or any other Delivery Service - exists that has the name "location" and
the Config File "remap.config". See apache/trafficcontrol#6691.